pyproject.toml
[tool.poetry]
name = "freela"
version = "0.1.0"
description = ""
authors = ["AriHenrique <arihenriquedev@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.31.0"
pandas = "^2.1.3"
jupyter = "^1.0.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
